requirements.txt
faker==25.8.0
pydantic==2.8.2
pyyaml==6.0.2
pytest==8.2.0
